requirements.txt
flask
flask_sqlalchemy
requests
boto3
python-dotenv